--- src/streams/files/fileToWrite.txt
@@ -1,2 +1,0 @@
-sdfsf
-sdfsdf
