+++ src/streams/write.js
@@ -1,4 +1,4 @@
-import {pipeline,Transform} from 'stream/promises'
+import {pipeline} from 'stream/promises'
 import {join, dirname} from 'path'
 import {open} from 'fs/promises'
 import {fileURLToPath} from 'url';
@@ -6,10 +6,6 @@ import {fileURLToPath} from 'url';
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 const filePath = join(__dirname, 'files', 'fileToWrite.txt');
-
-
-// write.js - implement function that writes process.stdin data into file fileToWrite.txt content using Writable Stream
-
 
 export const write = async () => {
   const ws = (await open(filePath, 'w')).createWriteStream();
